Added an option to use a timeout to start gestures
Mouse movement within this timeout of pressing the button binding will be interpreted as a gesture. This is mainly intended to allow the use of touchpads.

Unfortunately, this means adding a delay for any click with the button configured for gestures. Currently, it is not possible to distinguish between events coming from different pointer devices, so this will apply to all of them. Also, there is no way to know when the user lifts their finger from the touchpad, so the gesture is also ended with the timeout.