plugin: do not grab input when the binding is pressed
When we grab input, Wayfire will close all popups. This currently makes it impossible to right-click on popups if WStroke is active. We can use an alternative: we do not use an activator binding, but just listen to raw input events and activate based on these. This means that the surface under the mouse does not lose focus (and receives mouse movement) until a gesture is detected.